requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL (production)
dj-database-url>=2.1.0  # Parse DATABASE_URL
# SQLite used by default for development

# WhatsApp Cloud API
requests>=2.31.0
httpx>=0.25.0  # Async HTTP client

# Environment & Configuration
python-dotenv>=1.0.0
python-decouple>=3.8

# Production Server
gunicorn>=21.2.0
uvicorn>=0.24.0  # ASGI server
whitenoise>=6.6.0  # Static file serving

# Utilities
celery>=5.3.0  # Background tasks (optional)
redis>=5.0.0   # Celery broker (optional)

# Testing
pytest>=7.4.0
pytest-django>=4.7.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
